Abacus: ignore already existing Stripe events
It happened before that we saved the webhook payload to our database successfully, however server returned 503 because something got broken along the way. Later when Stripe retried to send the payload, our DB failed on "unique constraint violated" error because the event was already recorded.

This should fix this situation.

adeira-source-id: 68cb8cc8d5775b8fe4e4c020aeb35979b95c6380